Remove MonadFilter, and fix MonadCombineLaws
Fixes #1336

The `MonadCombine` typeclass unifies `Monad`, `FunctorFilter`, and
`MonoidK`, bringing three of the lineages together. This uses the
emptiness notion introduced in `MonoidK` and `flatMap` to implement
`mapFilter`, along with `unite`.

Given that `MonadCombine` exists, which is Cats equivalent of
`MonadPlus` there's no need for `MonadFilter` to exist.

The laws are migrated over from `MonadFilter` to `MonadCombine`, except
these laws would now check  identity laws and the consistencies against
the correct `F.empty` from `MonoidK`.
