+++ core/src/main/scala/cats/syntax/all.scala
@@ -25,7 +25,6 @@ trait AllSyntax
     with ListSyntax
     with MonadCombineSyntax
     with MonadErrorSyntax
-    with MonadFilterSyntax
     with MonoidSyntax
     with OptionSyntax
     with OrderSyntax
