--- docs/src/main/tut/monadfilter.md
@@ -1,9 +1,0 @@
----
-layout: default
-title:  "MonadFilter"
-section: "typeclasses"
-source: "core/src/main/scala/cats/MonadFilter.scala"
-scaladoc: "#cats.MonadFilter"
----
-# MonadFilter
-
