+++ tests/src/test/scala/cats/tests/ListWrapper.scala
@@ -104,8 +104,6 @@ object ListWrapper {
 
   def monoidK: MonoidK[ListWrapper] = monadCombine
 
-  def monadFilter: MonadFilter[ListWrapper] = monadCombine
-
   def alternative: Alternative[ListWrapper] = monadCombine
 
   def monoid[A]: Monoid[ListWrapper[A]] = monadCombine.algebra[A]
